+++ front/app/components/admin/ActionForm/AccessSections/IdMethodsModal/MethodBadges.tsx
